backend/requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
SQLAlchemy==2.0.48
passlib==1.7.4
bcrypt==3.2.2
PyJWT==2.12.1
python-dotenv==1.2.2
